Re-submission of gh-7087 with better file provenance (#7681)
* Split 3d animation into two examples.

* Rename example, preserve provenance.

* Apply updates to 3d anim examples.

* Change back to spectral layout.

* Update examples/3d_drawing/plot_3d_animation_walk.py


---------
